% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{addGraph,RCyjs-method}
\alias{addGraph,RCyjs-method}
\alias{addGraph}
\title{addGraph}
\usage{
\S4method{addGraph}{RCyjs}(obj, graph)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{graph}{a graphNEL}
}
\value{
nothing
}
\description{
\code{addGraph} send these nodes and edges (with attributes) to RCyjs for display
}
\details{
This version transmits a graph (nodes, edges and attributes) to the browser
by writing the data to a file, and sending that filename to be read in the
browser by javascript.
}
\examples{
if(interactive()){
  rcy <- RCyjs(title="rcyjs demo", graph=g)
  g <- simpleDemoGraph()
  setGraph(rcy, g)
  }

}
