% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{addGraphDirect,RCyjs-method}
\alias{addGraphDirect,RCyjs-method}
\alias{addGraphDirect}
\title{addGraphDirect}
\usage{
\S4method{addGraphDirect}{RCyjs}(obj, graph)
}
\arguments{
\item{p1}{obj RCyjs instance}

\item{p2}{graph a graphNEL}
}
\value{
nothing
}
\description{
\code{addDirectGraph} add graph, sending the data to the browser via the websocket
}
\details{
We usually prefer to send data to the browser by writing it to a file, and then
sending only the file's path to the browser.   This is the alternative.
It breaks down with large amounts of data, as the size approaches 1M.
if you like
}
\examples{
  rcy <- RCyjs(title="rcyjs demo", graph=g)
  g <- simpleDemoGraph()
  setGraphDirect(rcy, g)

}
