% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{clearSelection,RCyjs-method}
\alias{clearSelection,RCyjs-method}
\alias{clearSelection}
\title{clearSelection}
\usage{
\S4method{clearSelection}{RCyjs}(obj, which = "both")
}
\arguments{
\item{obj}{an RCyjs object}

\item{which}{a character string:  "both" (the default), "nodes" or "edges"}
}
\value{
no return value
}
\description{
\code{clearSelection} deselect all selected nodes, all selected edges, or both
}
\examples{
if(interactive()){
   rcy <- RCyjs(title="rcyjs demo", graph=simpleDemoGraph())
   selectNodes(rcy, c("A", "B"))
   clearSelection(rcy)
   }

}
