% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createTestGraph}
\alias{createTestGraph}
\title{createTestGraph}
\usage{
createTestGraph(nodeCount, edgeCount)
}
\arguments{
\item{nodeCount}{1 or more}

\item{edgeCount}{0 or more}
}
\value{
a graphNEL with \code{nodeCount} nodes and \code{edgeCount} edges
}
\description{
\code{createTestGraph} With as many nodes and edges as you wish, but neither edge nor node attributes.
}
\examples{
  g <- createTestGraph(5, 3)

}
