% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{deleteGraph,RCyjs-method}
\alias{deleteGraph,RCyjs-method}
\alias{deleteGraph}
\title{deleteGraph}
\usage{
\S4method{deleteGraph}{RCyjs}(obj)
}
\arguments{
\item{obj}{RCyjs instance}
}
\value{
nothing
}
\description{
\code{deleteGraph} Remove all nodes and edges, the elements of the current graph.
}
\details{
This method will remove any previous graph in the browser
}
\examples{
if(interactive()){
  sampleGraph <- simpleDemoGraph()
  rcy <- RCyjs(title="rcyjs demo", graph=sampleGraph)
  deletetGraph(rcy)
  }

}
\seealso{
\code{\link{addGraph}} \code{\link{setGraph}}
}
