% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{deleteSelectedNodes,RCyjs-method}
\alias{deleteSelectedNodes,RCyjs-method}
\alias{deleteSelectedNodes}
\title{deleteSelectedNodes}
\usage{
\S4method{deleteSelectedNodes}{RCyjs}(obj)
}
\arguments{
\item{obj}{an RCyjs instance}
}
\value{
explain what the method returns
}
\description{
\code{deleteSelectedNodes} put somewhat more detailed description here
}
\details{
multi-line description goes here with
continuations on subsequent lines
if you like
}
\examples{
if(interactive()){
   g <- simpleDemoGraph()
   rcy <- RCyjs(title="rcyjs demo", graph=g)
   target <- nodes(g)[1]
   selectNodes(rcy, target)
   deleteSelectedNodes(rcy)
   }

}
