% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\name{eda}
\alias{eda}
\title{eda}
\usage{
eda(graph, edge.attribute.name)
}
\arguments{
\item{g}{a graphNEL}
}
\value{
character strings, the names of the unique edge attribute categories on the graph
}
\description{
\code{eda} retrieve the node/attribute-value pairs, for the specified node attribute category
}
\examples{
  g <- simpleDemoGraph()
  edaNames(g)   # discover the attribute category names
  eda(g, "edgeType")
  eda(g, "score")

}
