% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\name{edaNames}
\alias{edaNames}
\title{edaNames}
\usage{
edaNames(graph)
}
\arguments{
\item{graph}{a graphNEL}
}
\value{
character strings, the names of the unique edge attribute categories on the graph
}
\description{
\code{edaNames} the names of the unique edge attribute categories on the graph (not their values)
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  edaNames(g)
  }

}
