% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{fit,RCyjs-method}
\alias{fit,RCyjs-method}
\alias{fit}
\title{fit}
\usage{
\S4method{fit}{RCyjs}(obj, padding = 30)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{padding}{numeric, in pixels}
}
\value{
explain what the method returns
}
\description{
\code{fit} put somewhat more detailed description here
}
\details{
multi-line description goes here with
continuations on subsequent lines
if you like
}
\examples{
  x <- 3 + 2

}
