% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{getJSON,RCyjs-method}
\alias{getJSON,RCyjs-method}
\alias{getJSON}
\title{getJSON}
\usage{
\S4method{getJSON}{RCyjs}(obj)
}
\arguments{
\item{obj}{an RCyjs instance}
}
\value{
a JSON string
}
\description{
\code{getJSON} a JSON string from the browser, describing the graph in cytoscape.js terms
}
\examples{
if(interactive()){
   sampleGraph <- simpleDemoGraph()
   rcy <- RCyjs(title="getJSON", graph=sampleGraph)
   s <- getJSON(rcy)
   s.asList <- fromJSON(s)  # easier to inspect if you wish toa
   }

}
