% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{getLayoutStrategies,RCyjs-method}
\alias{getLayoutStrategies,RCyjs-method}
\alias{getLayoutStrategies}
\title{getLayoutStrategies}
\usage{
\S4method{getLayoutStrategies}{RCyjs}(obj)
}
\arguments{
\item{obj}{an RCyjs instance}
}
\value{
a list of character strings
}
\description{
\code{getLayoutStrategies} return a list of those currently offered
}
\examples{


}
