% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{getNodeCount,RCyjs-method}
\alias{getNodeCount,RCyjs-method}
\alias{getNodeCount}
\title{getNodeCount}
\usage{
\S4method{getNodeCount}{RCyjs}(obj)
}
\arguments{
\item{obj}{RCyjs instance}
}
\value{
numeric count
}
\description{
\code{getNodeCount} the number of nodes in the current cytoscape.js graph
}
\examples{
if(interactive()){
   rcy <- RCyjs(title="rcyjs demo", graph=simpleDemoGraph())
   getNodeCount(rcy)
   }

}
