% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{getNodes,RCyjs-method}
\alias{getNodes,RCyjs-method}
\alias{getNodes}
\title{getNodes}
\usage{
\S4method{getNodes}{RCyjs}(obj, which = "all")
}
\arguments{
\item{obj}{an RCyjs instance}

\item{which}{a character string, either "all", "visible" or "hidden"}
}
\value{
a data.frame with at least and "id" column
}
\description{
\code{getNodes} returns a data.frame, one row per node, providing id and (if present) name and
label columns
}
\details{
Every node is guaranteed to have an "id" attribute.  Becuase "name" and "label" are commonly
used as well, they are returned as columns in the data.frame if present
}
\examples{
  rcy <- demo()

}
