% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{getPosition,RCyjs-method}
\alias{getPosition,RCyjs-method}
\alias{}
\title{getPosition}
\usage{
\S4method{getPosition}{RCyjs}(obj, nodeIDs = NA)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{nodeIDs}{a vector of character strings, default NA}
}
\value{
a data.frame with "id", "x" and "y" columns
}
\description{
\code{getPosition} for all or specified nodes
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="getPosition", graph=g)
  layout(rcy, "cose")
  tbl.pos <- getPosition(rcy)
  tbl.posA <- getPosition(rcy, "A")
  }

}
\seealso{
\code{\link{setPosition}}
}
