% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{getSelectedNodes,RCyjs-method}
\alias{getSelectedNodes,RCyjs-method}
\alias{getSelectedNodes}
\title{getSelectedNodes}
\usage{
\S4method{getSelectedNodes}{RCyjs}(obj)
}
\arguments{
\item{obj}{an RCyjs instance}
}
\value{
a data.frame with (at least) an id column
}
\description{
\code{getSelectedNodes} get the selected nodes
}
\examples{
if(interactive()){
   rcy <- RCyjs(title="rcyjs demo", graph=simpleDemoGraph())
   nodes.to.select <- getNodes(rcy)$id
   selectNodes(rcy, nodes.to.select)
   }

}
