% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{getSupportedEdgeDecoratorShapes,RCyjs-method}
\alias{getSupportedEdgeDecoratorShapes,RCyjs-method}
\alias{getSupportedEdgeDecoratorShapes}
\title{getSupportedEdgeDecoratorShapes}
\usage{
\S4method{getSupportedEdgeDecoratorShapes}{RCyjs}(obj)
}
\arguments{
\item{obj}{an RCyjs instance}
}
\value{
a list of character strings
}
\description{
\code{getSupportedEdgeDecoratorShapes} return a list of those currently offered
}
\examples{
if(interactive()){
  g <- createTestGraph(nodeCount=20, edgeCount=20)
  rcy <- RCyjs(title="shapes", graph=g)
  shapes <- getSupportedEdgeDecoratorShapes(rcy)
  }



}
