% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{getSupportedNodeShapes,RCyjs-method}
\alias{getSupportedNodeShapes,RCyjs-method}
\alias{getSupportedNodeShapes}
\title{getSupportedNodeShapes}
\usage{
\S4method{getSupportedNodeShapes}{RCyjs}(obj)
}
\arguments{
\item{obj}{an RCyjs instance}
}
\value{
a list of character strings
}
\description{
\code{getSupportedNodeShapes} return a list of those currently offered
}
\examples{
if(interactive()){
  g <- createTestGraph(nodeCount=20, edgeCount=20)
  rcy <- RCyjs(title="shapes", graph=g)
  shapes <- getSupportedNodeShapes(rcy)
  }

}
