% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{getZoom,RCyjs-method}
\alias{getZoom,RCyjs-method}
\alias{getZoom}
\title{getZoom}
\usage{
\S4method{getZoom}{RCyjs}(obj)
}
\arguments{
\item{obj}{an RCyjs instance}
}
\value{
no return value
}
\description{
\code{getZoom} learn the zoom level of the current display
}
\examples{
if(interactive()){
   rcy <- RCyjs(title="rcyjs demo", graph=simpleDemoGraph())
   getZoom(rcy)
   Sys.sleep(1)
   setZoom(rcy, 5)
   getZoom(rcy)
   }

}
