% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsToJSON.R
\name{graphNELtoJSON.string}
\alias{graphNELtoJSON.string}
\title{Convert R graphNEL object to cytoscape.js JSON.}
\usage{
graphNELtoJSON.string(g)
}
\arguments{
\item{g}{a graphNEL}
}
\description{
Convert R graphNEL object to cytoscape.js JSON.
}
\examples{
\dontrun{
  g.json <- graphNELtoJSON.string(graphNEL())
}

}
