% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{hAlign,RCyjs-method}
\alias{hAlign,RCyjs-method}
\alias{hAlign}
\title{hAlign}
\usage{
\S4method{hAlign}{RCyjs}(obj)
}
\arguments{
\item{obj}{an RCyjs instance}
}
\value{
no return value
}
\description{
\code{hAlign} horizontally align selected nodes
}
\details{
The shared y coordinate will be the mean of the y coordinates of selected nodes.
The x coordinates are preserved.
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="rcyjs demo", graph=g)
  layout(rcy, "cose")
  selectNodes(rcy, nodes(g)[1:2])
  hAlign(rcy)
  }


}
