% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{hideAllEdges,RCyjs-method}
\alias{hideAllEdges,RCyjs-method}
\alias{hideAllEdges}
\title{hideAllEdges}
\usage{
\S4method{hideAllEdges}{RCyjs}(obj)
}
\arguments{
\item{obj}{an RCyjs instance}
}
\value{
no return value
}
\description{
\code{hideAllEdges}
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="rcyjs demo", graph=g)
  layout(rcy, "cose")
  hideAllEdges()
  showAll(rcy, "edges")
  }
}
