% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{hideSelectedNodes,RCyjs-method}
\alias{hideSelectedNodes,RCyjs-method}
\alias{hideSelectedNodes}
\title{hideSelectedNodes}
\usage{
\S4method{hideSelectedNodes}{RCyjs}(obj)
}
\arguments{
\item{obj}{an RCyjs instance}
}
\value{
no return value
}
\description{
\code{hideSelectedNodes} hide selected nodes from view
}
\details{
The hidden nodes are not deleted from the graph
}
\examples{
if(interactive()){
   g <- simpleDemoGraph()
   rcy <- RCyjs(title="rcyjs demo", graph=g)
   target <- nodes(g)[1]
   selectNodes(rcy, target)
   hideSelectedNodes(rcy)
   getNodes(rcy, "hidden")
   getNodes(rcy, "visible")
   showAll(rcy, which="nodes")
   }

}
\seealso{
\code{\link{showAll}}
}
