% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{invertNodeSelection,RCyjs-method}
\alias{invertNodeSelection,RCyjs-method}
\alias{invertNodeSelection}
\title{invertNodeSelection}
\usage{
\S4method{invertNodeSelection}{RCyjs}(obj)
}
\arguments{
\item{obj}{an RCyjs instance}
}
\value{
no return value
}
\description{
\code{invertNodeSelection} deselect all selected nodes, select all previously unselected nodes
}
\examples{
if(interactive()){
   g <- simpleDemoGraph()
   rcy <- RCyjs(title="rcyjs demo", graph=g)
   target <- nodes(g)[1]
   selectNodes(rcy, target)
   invertNodeSelection(rcy)
   }

}
