% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{layout,RCyjs-method}
\alias{layout,RCyjs-method}
\alias{layout}
\title{layout}
\usage{
\S4method{layout}{RCyjs}(obj, strategy = "random")
}
\arguments{
\item{obj}{an RCyjs instance}

\item{strategy}{a character string, one of the supported algorithms}
}
\value{
explain what the method returns
}
\description{
\code{layout} apply a layout algorithm to the current grap
}
\examples{
if(interactive()){
  g <- createTestGraph(nodeCount=20, edgeCount=20)
  rcy <- RCyjs(title="layouts", graph=g)
  strategies <- getLayoutStrategies(rcy)
  for(strategy in stategies){
     layout(rcy, strategy)
     Sys.sleep(1)
     }
  }

}
\seealso{
\code{\link{getLayoutStrategies}}
}
