% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{layoutSelectionInGrid,RCyjs-method}
\alias{layoutSelectionInGrid,RCyjs-method}
\alias{layoutSelectionInGrid}
\title{layoutSelectionInGrid}
\usage{
\S4method{layoutSelectionInGrid}{RCyjs}(obj, x, y, w, h)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{x}{numeric}

\item{y}{numeric}

\item{w}{numeric}

\item{h}{numeric}
}
\value{
explain what the method returns
}
\description{
\code{layoutSelectionInGrid} put somewhat more detailed description here
}
\details{
multi-line description goes here with
continuations on subsequent lines
if you like
}
\examples{
  x <- 3 + 2

}
