% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{layoutSelectionInGridInferAnchor,RCyjs-method}
\alias{layoutSelectionInGridInferAnchor,RCyjs-method}
\alias{layoutSelectionInGridInferAnchor}
\title{layoutSelectionInGridInferAnchor}
\usage{
\S4method{layoutSelectionInGridInferAnchor}{RCyjs}(obj, w, h)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{x}{numeric}

\item{y}{numeric}
}
\value{
explain what the method returns
}
\description{
\code{layoutSelectionInGridInferAnchor} put somewhat more detailed description here
}
\details{
multi-line description goes here with
continuations on subsequent lines
if you like
}
\examples{
  x <- 3 + 2

}
