% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{loadStyleFile,RCyjs-method}
\alias{loadStyleFile,RCyjs-method}
\alias{loadStyleFile}
\title{loadStyleFile}
\usage{
\S4method{loadStyleFile}{RCyjs}(obj, filename)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{filename}{contains json in the proper cytoscape.js format}
}
\value{
nothing
}
\description{
\code{loadStyleFile} load a named JSON cytoscape.js style file into the browser
}
\examples{
if(interactive()){
  rcy <- demo()
  filename <- system.file(package="RCyjs", "extdata", "sampleStyle1.js");
  loadStyleFile(rcy, filename)
  }

}
\references{
\url{https://js.cytoscape.org/#style}

Though we provide access to individual styling rules (see below) we often find
it convenient to express all aspects of a visual style in a single JSON file
}
