% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{hideEdges,RCyjs-method}
\alias{hideEdges,RCyjs-method}
\alias{methodname}
\alias{showEdges,RCyjs-method}
\alias{methodname}
\alias{getZoom,RCyjs-method}
\alias{methodname}
\alias{setZoom,RCyjs-method}
\alias{methodname}
\alias{setBackgroundColor,RCyjs-method}
\alias{methodname}
\alias{vAlign,RCyjs-method}
\alias{methodname}
\alias{hAlign,RCyjs-method}
\alias{methodname}
\title{title}
\usage{
\S4method{hideEdges}{RCyjs}(obj, edgeType)

\S4method{showEdges}{RCyjs}(obj, edgeType)

\S4method{getZoom}{RCyjs}(obj)

\S4method{setZoom}{RCyjs}(obj, newValue)

\S4method{setBackgroundColor}{RCyjs}(obj, newValue)

\S4method{vAlign}{RCyjs}(obj)

\S4method{hAlign}{RCyjs}(obj)
}
\arguments{
\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}

\item{p1}{some text}

\item{p2}{some text}

\item{p3}{some text}
}
\value{
explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns

explain what the method returns
}
\description{
\code{methodName} put somewhat more detailed description here

\code{methodName} put somewhat more detailed description here

\code{methodName} put somewhat more detailed description here

\code{methodName} put somewhat more detailed description here

\code{methodName} put somewhat more detailed description here

\code{methodName} put somewhat more detailed description here

\code{methodName} put somewhat more detailed description here
}
\details{
multi-line description goes here with
continuations on subsequent lines
if you like

multi-line description goes here with
continuations on subsequent lines
if you like

multi-line description goes here with
continuations on subsequent lines
if you like

multi-line description goes here with
continuations on subsequent lines
if you like

multi-line description goes here with
continuations on subsequent lines
if you like

multi-line description goes here with
continuations on subsequent lines
if you like

multi-line description goes here with
continuations on subsequent lines
if you like
}
\examples{
  x <- 3 + 2

  x <- 3 + 2

  x <- 3 + 2

  x <- 3 + 2

  x <- 3 + 2

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2
title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

title

\\code{methodName} put somewhat more detailed description here

multi-line description goes here with
continuations on subsequent lines
if you like

  x <- 3 + 2

  x <- 3 + 2

}
