% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\name{noa}
\alias{noa}
\title{noa}
\usage{
noa(graph, node.attribute.name)
}
\arguments{
\item{g}{a graphNEL}
}
\value{
character strings, the names of the unique edge attribute categories on the graph
}
\description{
\code{noa} retrieve the node/attribute-value pairs, for the specified node attribute category
}
\examples{
  g <- simpleDemoGraph()
  noa(g, "lfc")

}
