% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\name{noaNames}
\alias{noaNames}
\title{noaNames}
\usage{
noaNames(graph)
}
\arguments{
\item{graph}{a graphNEL}
}
\value{
character strings, the names of the unique node attribute categories on the graph
}
\description{
\code{noaNames} the names of the unique node attribute categories on the graph (not their values)
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  noaNames(g)
  }

}
