% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{redraw,RCyjs-method}
\alias{redraw,RCyjs-method}
\alias{redraw}
\title{redraw}
\usage{
\S4method{redraw}{RCyjs}(obj)
}
\arguments{
\item{obj}{an RCyjs instance}
}
\value{
explain what the method returns
}
\description{
\code{redraw} put somewhat more detailed description here
}
\details{
multi-line description goes here with
continuations on subsequent lines
if you like
}
\examples{
  x <- 3 + 2

}
