% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{restoreLayout,RCyjs-method}
\alias{restoreLayout,RCyjs-method}
\alias{restoreLayout}
\title{restoreLayout}
\usage{
\S4method{restoreLayout}{RCyjs}(obj, filename = "layout.RData")
}
\arguments{
\item{obj}{an RCyjs instance}

\item{filename}{a character string, default "layout.RData"}
}
\value{
no return value
}
\description{
\code{restoreLayout} restore a previously-saved layout
}
\examples{
if(interactive()){
   rcy <- RCyjs(title="rcyjs demo", graph=simpleDemoGraph())
   layout(rcy, "grid")
   saveLayout(rcy, filename="gridLayout.RData")
   layout(rcy, "circle")
   restoreLayout(rcy, "gridLayout.RData")
   }

}
\seealso{
\code{\link{saveLayout}}
}
