% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{saveLayout,RCyjs-method}
\alias{saveLayout,RCyjs-method}
\alias{saveLayout}
\title{saveLayout}
\usage{
\S4method{saveLayout}{RCyjs}(obj, filename = "layout.RData")
}
\arguments{
\item{obj}{a RCyjs instance}

\item{filename}{"layout.RData" by default}
}
\value{
no return value
}
\description{
\code{saveLayout} to a named file
}
\details{
All node positions are saved to a functionally opaque RData object,
in a file whose name you supply.  These files are used by
restoreLayout.
}
\examples{
if(interactive()){
   rcy <- RCyjs(title="rcyjs demo", graph=simpleDemoGraph())
   layout(rcy, "grid")
   saveLayout(rcy, filename="gridLayout.RData")
   layout(rcy, "circle")
   restoreLayout(rcy, "gridLayout.RData")
   }

}
\seealso{
\code{\link{restoreLayout}}
}
