% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{savePNG,RCyjs-method}
\alias{savePNG,RCyjs-method}
\alias{savePNG}
\title{savePNG}
\usage{
\S4method{savePNG}{RCyjs}(obj, filename)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{filename}{a character string}
}
\value{
no return value
}
\description{
\code{savePNG} write current cytoscape view, at current resolution, to a PNG file.
}
\examples{
if(interactive()){
   rcy <- RCyjs(title="layouts", graph=createTestGraph(nodeCount=20, edgeCount=20))
   style.filename <- system.file(package="RCyjs", "extdata", "sampleStyle1.js");
   loadStyleFile(rcy, style.filename)
   layout(rcy, "cose")
   fit(rcy)
   filename <- tempfile(fileext=".png")
   savePNG(rcy, filename)
   }
}
