% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{selectFirstNeighborsOfSelectedNodes,RCyjs-method}
\alias{selectFirstNeighborsOfSelectedNodes,RCyjs-method}
\alias{selectFirstNeighborsOfSelectedNodes}
\title{selectFirstNeighborsOfSelectedNodes}
\usage{
\S4method{selectFirstNeighborsOfSelectedNodes}{RCyjs}(obj)
}
\arguments{
\item{obj}{an RCyjs instance}
}
\value{
no return value
}
\description{
\code{selectFirstNeighborsOfSelectedNodes}
}
\examples{
if(interactive()){
   rcy <- RCyjs(title="rcyjs demo", graph=simpleDemoGraph())
   selectNodes(rcy, "A")
   getSelectedNodes(rcy)  # just one
   selectFirstNeighborsOfSelectedNodes()
   getSelectedNodes(rcy)  # now three
   }

}
