% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{selectNodes,RCyjs-method}
\alias{selectNodes,RCyjs-method}
\alias{selectNodes}
\title{selectNodes}
\usage{
\S4method{selectNodes}{RCyjs}(obj, nodeIDs)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{nodeIDs}{character strings}
}
\value{
no return value
}
\description{
\code{selectNodes} put somewhat more detailed description here
}
\details{
multi-line description goes here with
continuations on subsequent lines
if you like
}
\examples{
  x <- 3 + 2

}
