% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setBackgroundColor,RCyjs-method}
\alias{setBackgroundColor,RCyjs-method}
\alias{setBackgroundColor}
\title{setBackgroundColor}
\usage{
\S4method{setBackgroundColor}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a character string, any valid CSS color}
}
\value{
no return value
}
\description{
\code{setBackgroundColor} of the entire cytoscape.js div
}
\examples{
if(interactive()){
   rcy <- RCyjs(title="rcyjs demo", graph=simpleDemoGraph())
   setBackgroundColor(rcy, "lightblue")
   }

}
