% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setDefaultEdgeColor,RCyjs-method}
\alias{setDefaultEdgeColor,RCyjs-method}
\alias{setDefaultEdgeColor}
\title{setDefaultEdgeColor}
\usage{
\S4method{setDefaultEdgeColor}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a character string, any valid CSS color}
}
\value{
no value returned
}
\description{
\code{setDefaultEdgeColor}
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setDefaultNodeColor", graph=g)
  layout(rcy, "cose")
  setDefaultNodeFontColor(rcy, "red")
  }

}
