% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setDefaultEdgeLineColor,RCyjs-method}
\alias{setDefaultEdgeLineColor,RCyjs-method}
\alias{setDefaultEdgeLineColor}
\title{setDefaultEdgeLineColor}
\usage{
\S4method{setDefaultEdgeLineColor}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a character string, and valid CSS color}
}
\value{
no value returned
}
\description{
\code{setDefaultEdgeLineColor}
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setDefaultEdgeLineColor", graph=g)
  layout(rcy, "cose")
  setDefaultEdgeLineColor(rcy, "red")
  }

}
