% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setDefaultEdgeLineStyle,RCyjs-method}
\alias{setDefaultEdgeLineStyle,RCyjs-method}
\alias{setDefaultEdgeLineStyle}
\title{setDefaultEdgeLineStyle}
\usage{
\S4method{setDefaultEdgeLineStyle}{RCyjs}(obj, newValue = c("solid", "dotted",
  "dashed"))
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a character string, one of "solid", "dotted", or "dashed"}
}
\value{
no value returned
}
\description{
\code{setDefaultEdgeLineStyle} put somewhat more detailed description here
}
\details{
multi-line description goes here with
continuations on subsequent lines
if you like
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setDefaultEdgeLineStyle", graph=g)
  layout(rcy, "cose")
  setDefaultEdgeLineColor(rcy, "dashed")
  }

}
