% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setDefaultEdgeSourceArrowColor,RCyjs-method}
\alias{setDefaultEdgeSourceArrowColor,RCyjs-method}
\alias{setDefaultEdgeSourceArrowColor}
\title{setDefaultEdgeSourceArrowColor}
\usage{
\S4method{setDefaultEdgeSourceArrowColor}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a character string, and valid CSS color}
}
\value{
no value returned
}
\description{
\code{setDefaultEdgeSourceArrowColor}
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setDefaultEdgeSourceArrowColor", graph=g)
  layout(rcy, "cose")
  setDefaultEdgeSourceArrowColor(rcy, "red")
  }

}
