% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setDefaultEdgeTargetArrowColor,RCyjs-method}
\alias{setDefaultEdgeTargetArrowColor,RCyjs-method}
\alias{setDefaultEdgeTargetArrowColor}
\title{setDefaultEdgeTargetArrowColor}
\usage{
\S4method{setDefaultEdgeTargetArrowColor}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a character string, and valid CSS color}
}
\value{
no value returned
}
\description{
\code{setDefaultEdgeTargetArrowColor}
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setDefaultEdgeTargetArrowColor", graph=g)
  layout(rcy, "cose")
  setDefaultEdgeTargetArrowColor(rcy, "red")
  }

}
