% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setDefaultEdgeTargetArrowShape,RCyjs-method}
\alias{setDefaultEdgeTargetArrowShape,RCyjs-method}
\alias{setDefaultEdgeTargetArrowShape}
\title{setDefaultEdgeTargetArrowShape}
\usage{
\S4method{setDefaultEdgeTargetArrowShape}{RCyjs}(obj, newValue = c("triangle",
  "triangle-tee", "triangle-cross", "triangle-backcurve", "vee", "tee",
  "square", "circle", "diamond", "none"))
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a character string, one of "triangle", "triangle-tee", "triangle-cross", "triangle-backcurve",
"vee", "tee", "square", "circle", "diamond", "none"}
}
\value{
no value returned
}
\description{
\code{setDefaultEdgeTargetArrowShape} put somewhat more detailed description here
}
\details{
multi-line description goes here with
continuations on subsequent lines
if you like
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setDefaultEdgeTargetArrowShape", graph=g)
  layout(rcy, "cose")
  setDefaultEdgeTargetArrowShape(rcy, "tee")
  }

}
