% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setDefaultEdgeWidth,RCyjs-method}
\alias{setDefaultEdgeWidth,RCyjs-method}
\alias{setDefaultEdgeWidth}
\title{setDefaultEdgeWidth}
\usage{
\S4method{setDefaultEdgeWidth}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{}
}
\value{
no value returned
}
\description{
\code{setDefaultEdgeWidth} in pixels
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setDefaultEdgeWidth", graph=g)
  layout(rcy, "cose")
  setDefaultEdgeWidth(rcy, "1")
  }

}
