% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setDefaultNodeBorderWidth,RCyjs-method}
\alias{setDefaultNodeBorderWidth,RCyjs-method}
\alias{setDefaultNodeBoderWidth}
\title{setDefaultNodeBoderWidth}
\usage{
\S4method{setDefaultNodeBorderWidth}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{numeric, in pixels}
}
\value{
no value returned
}
\description{
\code{setDefaultNodeBoderWidth} in pixels
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setDefaultNodeBorderWidth", graph=g)
  layout(rcy, "cose")
  setDefaultNodeBorderWidth(rcy, 2)
  }
}
