% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setDefaultNodeBorderColor,RCyjs-method}
\alias{setDefaultNodeBorderColor,RCyjs-method}
\alias{setDefaultNodeBorderColor}
\title{setDefaultNodeBorderColor}
\usage{
\S4method{setDefaultNodeBorderColor}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{any CSS color}
}
\value{
no value returned
}
\description{
\code{setDefaultNodeBorderColor} put somewhat more detailed description here
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setDefaultNodeBorderColor", graph=g)
  layout(rcy, "cose")
  setDefaultNodeBorderColor(rcy, "red")
  }

}
