% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setDefaultNodeColor,RCyjs-method}
\alias{setDefaultNodeColor,RCyjs-method}
\alias{setDefaultNodeColor}
\title{setDefaultNodeColor}
\usage{
\S4method{setDefaultNodeColor}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a character string, any valid CSS color name}
}
\value{
no value returned
}
\description{
\code{setDefaultNodeColor} put somewhat more detailed description here
}
\details{
multi-line description goes here with
continuations on subsequent lines
if you like
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setDefaultNodeColor", graph=g)
  layout(rcy, "cose")
  setDefaultNodeColor(rcy, "lightblue")
  }

}
