% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setDefaultNodeFontColor,RCyjs-method}
\alias{setDefaultNodeFontColor,RCyjs-method}
\alias{setDefaultNodeFontColor}
\title{setDefaultNodeFontColor}
\usage{
\S4method{setDefaultNodeFontColor}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{any CSS color}
}
\value{
no value returned
}
\description{
\code{setDefaultNodeFontColor}
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setDefaultNodeColor", graph=g)
  layout(rcy, "cose")
  setDefaultNodeFontColor(rcy, "red")
  }

}
