% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setDefaultNodeFontSize,RCyjs-method}
\alias{setDefaultNodeFontSize,RCyjs-method}
\alias{setDefaultNodeFontSize}
\title{setDefaultNodeFontSize}
\usage{
\S4method{setDefaultNodeFontSize}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{numeric, in points}
}
\value{
no value returned
}
\description{
\code{setDefaultNodeFontSize} put somewhat more detailed description here
}
\details{
multi-line description goes here with
continuations on subsequent lines
if you like
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setDefaultNodeFontSize", graph=g)
  layout(rcy, "cose")
  setDefaultNodeFontSize(rcy, 8)
  }

}
