% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setDefaultNodeHeight,RCyjs-method}
\alias{setDefaultNodeHeight,RCyjs-method}
\alias{setDefaultNodeHeight}
\title{setDefaultNodeHeight}
\usage{
\S4method{setDefaultNodeHeight}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a numeric, in pixels}
}
\value{
no value returned
}
\description{
\code{setDefaultNodeHeight} set all nodes to the same specifed width, in pixels
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setDefaultNodeWidth", graph=g)
  layout(rcy, "cose")
  setDefaultNodeHeight(rcy, 80)
  }

}
