% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setDefaultNodeShape,RCyjs-method}
\alias{setDefaultNodeShape,RCyjs-method}
\alias{setDefaultNodeShape}
\title{setDefaultNodeShape}
\usage{
\S4method{setDefaultNodeShape}{RCyjs}(obj, newValue = c("ellipse", "triangle",
  "rectangle", "roundrectangle", "bottomroundrectangle", "cutrectangle",
  "barrel", "rhomboid", "diamond", "pentagon", "hexagon", "concavehexagon",
  "heptagon", "octagon", "star", "tag", "vee"))
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a character string, one of "ellipse", "triangle", "rectangle", "roundrectangle",
"bottomroundrectangle","cutrectangle", "barrel",
"rhomboid", "diamond", "pentagon", "hexagon",
"concavehexagon", "heptagon", "octagon", "star", "tag", "vee"}
}
\value{
no value returned
}
\description{
\code{setDefaultNodeShape} change the shape of all nodes
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setDefaultNodeShape", graph=g)
  layout(rcy, "cose")
  setDefaultNodeShape(rcy, "barrel")
  }

}
