% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setDefaultNodeSize,RCyjs-method}
\alias{setDefaultNodeSize,RCyjs-method}
\alias{setDefaultNodeSize}
\title{setDefaultNodeSize}
\usage{
\S4method{setDefaultNodeSize}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a numeric, in pixels}
}
\value{
no value returned
}
\description{
\code{setDefaultNodeSize} set all nodes to the same specifed size, in pixels
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setDefaultNodesSize", graph=g)
  layout(rcy, "cose")
  setDefaultNodeSize(rcy, 80)
  }

}
