% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setDefaultNodeWidth,RCyjs-method}
\alias{setDefaultNodeWidth,RCyjs-method}
\alias{setDefaultNodeWidth}
\title{setDefaultNodeWidth}
\usage{
\S4method{setDefaultNodeWidth}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a numeric, in pixels}
}
\value{
no value returned
}
\description{
\code{setDefaultNodeWidth} set all nodes to the same specifed width, in pixels
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setDefaultNodesWidth", graph=g)
  layout(rcy, "cose")
  setDefaultNodeWidth(rcy, 80)
  }

}
