% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setDefaultStyle,RCyjs-method}
\alias{setDefaultStyle,RCyjs-method}
\alias{setDefaultStyle}
\title{setDefaultStyle}
\usage{
\S4method{setDefaultStyle}{RCyjs}(obj)
}
\arguments{
\item{obj}{an RCyjs instance}
}
\value{
no value returned
}
\description{
\code{setDefaultStyle} use some sensible rendering options for all elements of the graph
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setDefaultNodesSize", graph=g)
  layout(rcy, "cose")
  setDefaultStyle(rcy)
  }

}
