% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setEdgeAttributes,RCyjs-method}
\alias{setEdgeAttributes,RCyjs-method}
\alias{setEdgeAttributes}
\title{setEdgeAttributes}
\usage{
\S4method{setEdgeAttributes}{RCyjs}(obj, attribute, sourceNodes, targetNodes,
  edgeTypes, values)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{sourceNodes}{vector of character strings}

\item{targetNodes}{vector of character strings}

\item{edgeTypes}{vector of character strings}

\item{values}{vector of character strings}
}
\value{
no return value
}
\description{
\code{setEdgeAttributes} on the graph in the browse
}
\details{
Edges are specified by sourceNode/targetNode/edgeType triples.
}
\examples{
  x <- 3 + 2

}
