% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setEdgeColorRule,RCyjs-method}
\alias{setEdgeColorRule,RCyjs-method}
\alias{setEdgeColorRule}
\title{setEdgeColorRule}
\usage{
\S4method{setEdgeColorRule}{RCyjs}(obj, attribute, control.points, colors,
  mode = c("interpolate", "lookup"))
}
\arguments{
\item{obj}{an RCyjs instance}

\item{attribute}{a character string, the edge attribute category whose value controls color}

\item{control.points}{a list of all possible values of the attribute}

\item{colors}{the corresponding edge color, one specified for each of the control.points}

\item{mode}{}
}
\value{
no return value
}
\description{
\code{setEdgeColorRule} control edge color via values of the specified attribute
}
\details{
for interpolate mode, in which the edge attribute should be a continusously varying numerical quantity
in-between colors are calculated for in-between values.
for lookup mode, in which the edge attribute is a discrete string variable, simple color lookup is performed.
}
\examples{
  x <- 3 + 2

}
