% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setEdgeSourceArrowShapeRule,RCyjs-method}
\alias{setEdgeSourceArrowShapeRule,RCyjs-method}
\alias{setEdgeSourceArrowShapeRule}
\title{setEdgeSourceArrowShapeRule}
\usage{
\S4method{setEdgeSourceArrowShapeRule}{RCyjs}(obj, attribute, control.points,
  shapes = c("triangle", "triangle-tee", "triangle-cross",
  "triangle-backcurve", "vee", "tee", "square", "circle", "diamond", "none"))
}
\arguments{
\item{obj}{an RCyjs instance}

\item{attribute}{a character string, the edgeSourceArrow attribute category whose value controls shape}

\item{control.points}{a list of all possible values of the attribute}

\item{shapes}{the corresponding edgeSourceArrow shape, one specified for each of the control.points}
}
\value{
no return value
}
\description{
\code{setEdgeSourceArrowShapeRule} control edgeSourceArrow shape via values of the specified attribute
}
\details{
always operates in lookup mode: interpolating between shapes is not possible here.
}
\examples{
  x <- 3 + 2

}
