% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setEdgeTargetArrowColorRule,RCyjs-method}
\alias{setEdgeTargetArrowColorRule,RCyjs-method}
\alias{setEdgeTargetArrowColorRule}
\title{setEdgeTargetArrowColorRule}
\usage{
\S4method{setEdgeTargetArrowColorRule}{RCyjs}(obj, attribute, control.points,
  colors, mode = c("interpolate", "lookup"))
}
\arguments{
\item{obj}{an RCyjs instance}

\item{attribute}{a character string, the edgeSourceArrow attribute category whose value controls shape}

\item{control.points}{a list of all possible values of the attribute}

\item{shapes}{the corresponding edgeSourceArrow shape, one specified for each of the control.points}
}
\value{
no return value
}
\description{
\code{setEdgeTargetArrowColorRule} control edgeSourceArrow shape via values of the specified attribute
}
\details{
for interpolate mode, in which the edge attribute should be a continusously varying numerical quantity
in-between colors are calculated for in-between values.
for lookup mode, in which the edge attribute is a discrete string variable, simple color lookup is performed.
}
\examples{
  x <- 3 + 2

}
