% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setEdgeTargetArrowShapeRule,RCyjs-method}
\alias{setEdgeTargetArrowShapeRule,RCyjs-method}
\alias{setEdgeTargetArrowShapeRule}
\title{setEdgeTargetArrowShapeRule}
\usage{
\S4method{setEdgeTargetArrowShapeRule}{RCyjs}(obj, attribute, control.points,
  shapes = c("triangle", "triangle-tee", "triangle-cross",
  "triangle-backcurve", "vee", "tee", "square", "circle", "diamond", "none"))
}
\arguments{
\item{obj}{an RCyjs instance}

\item{attribute}{a character string, the edgeTargetArrow attribute category whose value controls shape}

\item{control.points}{a list of all possible values of the attribute}

\item{shapes}{the corresponding edgeTargetArrow shape, one specified for each of the control.points}
}
\value{
no return value
}
\description{
\code{setEdgeTargetArrowShapeRule} control edgeTargetArrow shape via values of the specified attribute
}
\details{
always operates in lookup mode: interpolating between shapes is not possible here.
}
\examples{
  x <- 3 + 2

}
