% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setEdgeWidthRule,RCyjs-method}
\alias{setEdgeWidthRule,RCyjs-method}
\alias{setEdgeWidthRule}
\title{setEdgeWidthRule}
\usage{
\S4method{setEdgeWidthRule}{RCyjs}(obj, attribute, control.points, widths,
  mode = c("interpolate", "lookup"))
}
\arguments{
\item{obj}{an RCyjs instance}

\item{attribute}{a character string, the node attribute category whose value controls size}

\item{control.points}{a list of values of the attribute}

\item{thickness}{the corresponding line thickness, one specified for each of the control.points}
}
\value{
no return value
}
\description{
\code{setEdgeWidthRule} control node size via values of the specified attribute
}
\details{
actual node sizes are interpolated via the specified relationship of control.points node.sizes
}
\examples{
  x <- 3 + 2

}
