% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setGlobalEdgeColor,RCyjs-method}
\alias{setGlobalEdgeColor,RCyjs-method}
\alias{setGlobalEdgeColor}
\title{setGlobalEdgeColor}
\usage{
\S4method{setGlobalEdgeColor}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a character string, any valid CSS color}
}
\value{
no value returned
}
\description{
\code{setGlobalEdgeColor}
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setGlobalNodeColor", graph=g)
  layout(rcy, "cose")
  setGlobalNodeFontColor(rcy, "red")
  }

}
