% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setGlobalEdgeLineColor,RCyjs-method}
\alias{setGlobalEdgeLineColor,RCyjs-method}
\alias{setGlobalEdgeLineColor}
\title{setGlobalEdgeLineColor}
\usage{
\S4method{setGlobalEdgeLineColor}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a character string, and valid CSS color}
}
\value{
no value returned
}
\description{
\code{setGlobalEdgeLineColor}
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setGlobalEdgeLineColor", graph=g)
  layout(rcy, "cose")
  setGlobalEdgeLineColor(rcy, "red")
  }

}
