% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setGlobalEdgeSourceArrowColor,RCyjs-method}
\alias{setGlobalEdgeSourceArrowColor,RCyjs-method}
\alias{setGlobalEdgeSourceArrowColor}
\title{setGlobalEdgeSourceArrowColor}
\usage{
\S4method{setGlobalEdgeSourceArrowColor}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a character string, and valid CSS color}
}
\value{
no value returned
}
\description{
\code{setGlobalEdgeSourceArrowColor}
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setGlobalEdgeSourceArrowColor", graph=g)
  layout(rcy, "cose")
  setGlobalEdgeSourceArrowColor(rcy, "red")
  }

}
