% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setGlobalEdgeSourceArrowShape,RCyjs-method}
\alias{setGlobalEdgeSourceArrowShape,RCyjs-method}
\alias{setGlobalEdgeSourceArrowShape}
\title{setGlobalEdgeSourceArrowShape}
\usage{
\S4method{setGlobalEdgeSourceArrowShape}{RCyjs}(obj, newValue = c("triangle",
  "triangle-tee", "triangle-cross", "triangle-backcurve", "vee", "tee",
  "square", "circle", "diamond", "none"))
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a character string, one of "triangle", "triangle-tee", "triangle-cross", "triangle-backcurve",
"vee", "tee", "square", "circle", "diamond", "none"}
}
\value{
no value returned
}
\description{
\code{setGlobalEdgeSourceArrowShape} put somewhat more detailed description here
}
\details{
multi-line description goes here with
continuations on subsequent lines
if you like
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setGlobalEdgeSourceArrowShape", graph=g)
  layout(rcy, "cose")
  setGlobalEdgeSourceArrowShape(rcy, "tee")
  }

}
