% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setGlobalEdgeTargetArrowColor,RCyjs-method}
\alias{setGlobalEdgeTargetArrowColor,RCyjs-method}
\alias{setGlobalEdgeTargetArrowColor}
\title{setGlobalEdgeTargetArrowColor}
\usage{
\S4method{setGlobalEdgeTargetArrowColor}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a character string, and valid CSS color}
}
\value{
no value returned
}
\description{
\code{setGlobalEdgeTargetArrowColor}
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setGlobalEdgeTargetArrowColor", graph=g)
  layout(rcy, "cose")
  setGlobalEdgeTargetArrowColor(rcy, "red")
  }

}
