% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setGlobalEdgeWidth,RCyjs-method}
\alias{setGlobalEdgeWidth,RCyjs-method}
\alias{setGlobalEdgeWidth}
\title{setGlobalEdgeWidth}
\usage{
\S4method{setGlobalEdgeWidth}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a numeric}
}
\value{
no value returned
}
\description{
\code{setGlobalEdgeWidth} in pixels
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setGlobalEdgeWidth", graph=g)
  layout(rcy, "cose")
  setGlobalEdgeWidth(rcy, 1)
  }

}
