% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setGlobalNodeBorderColor,RCyjs-method}
\alias{setGlobalNodeBorderColor,RCyjs-method}
\alias{setGlobalNodeBorderColor}
\title{setGlobalNodeBorderColor}
\usage{
\S4method{setGlobalNodeBorderColor}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{any CSS color}
}
\value{
no value returned
}
\description{
\code{setGlobalNodeBorderColor} put somewhat more detailed description here
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setGlobalNodeBorderColor", graph=g)
  layout(rcy, "cose")
  setGlobalNodeBorderColor(rcy, "red")
  }

}
