% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setGlobalNodeBorderWidth,RCyjs-method}
\alias{setGlobalNodeBorderWidth,RCyjs-method}
\alias{setGlobalNodeBorderWidth}
\title{setGlobalNodeBorderWidth}
\usage{
\S4method{setGlobalNodeBorderWidth}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{numeric, in pixels}
}
\value{
no value returned
}
\description{
\code{setGlobalNodeBorderWidth} in pixels
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setGlobalNodeBorderWidth", graph=g)
  layout(rcy, "cose")
  setGlobalNodeBorderWidth(rcy, 2)
  }
}
