% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setGlobalNodeColor,RCyjs-method}
\alias{setGlobalNodeColor,RCyjs-method}
\alias{setGlobalNodeColor}
\title{setGlobalNodeColor}
\usage{
\S4method{setGlobalNodeColor}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{}
}
\value{
no value returned
}
\description{
\code{setGlobalNodeColor} put somewhat more detailed description here
}
\details{
multi-line description goes here with
continuations on subsequent lines
if you like
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setGlobalNodeColor", graph=g)
  layout(rcy, "cose")
  setGlobalNodeColor(rcy, "lightblue")
  }

}
