% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setGlobalNodeFontColor,RCyjs-method}
\alias{setGlobalNodeFontColor,RCyjs-method}
\alias{setGlobalNodeFontColor}
\title{setGlobalNodeFontColor}
\usage{
\S4method{setGlobalNodeFontColor}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{any CSS color}
}
\value{
no value returned
}
\description{
\code{setGlobalNodeFontColor}
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setGlobalNodeColor", graph=g)
  layout(rcy, "cose")
  setGlobalNodeFontColor(rcy, "red")
  }

}
