% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setGlobalNodeFontSize,RCyjs-method}
\alias{setGlobalNodeFontSize,RCyjs-method}
\alias{setGlobalNodeFontSize}
\title{setGlobalNodeFontSize}
\usage{
\S4method{setGlobalNodeFontSize}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{}
}
\value{
no value returned
}
\description{
\code{setGlobalNodeFontSize} put somewhat more detailed description here
}
\details{
multi-line description goes here with
continuations on subsequent lines
if you like
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setGlobalNodeFontSize", graph=g)
  layout(rcy, "cose")
  setGlobalNodeFontSize(rcy, 8)
  }

}
