% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setGlobalNodeHeight,RCyjs-method}
\alias{setGlobalNodeHeight,RCyjs-method}
\alias{setGlobalNodeHeight}
\title{setGlobalNodeHeight}
\usage{
\S4method{setGlobalNodeHeight}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a numeric, in pixels}
}
\value{
no value returned
}
\description{
\code{setGlobalNodeHeight} set all nodes to the same specifed width, in pixels
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setGlobalNodeHeight", graph=g)
  layout(rcy, "cose")
  setGlobalNodeHeight(rcy, 80)
  }

}
