% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setGlobalNodeShape,RCyjs-method}
\alias{setGlobalNodeShape,RCyjs-method}
\alias{setGlobalNodeShape}
\title{setGlobalNodeShape}
\usage{
\S4method{setGlobalNodeShape}{RCyjs}(obj, newValue = c("ellipse", "triangle",
  "rectangle", "roundrectangle", "bottomroundrectangle", "cutrectangle",
  "barrel", "rhomboid", "diamond", "pentagon", "hexagon", "concavehexagon",
  "heptagon", "octagon", "star", "tag", "vee"))
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a character string, one of "ellipse", "triangle", "rectangle", "roundrectangle",
"bottomroundrectangle","cutrectangle", "barrel",
"rhomboid", "diamond", "pentagon", "hexagon",
"concavehexagon", "heptagon", "octagon", "star", "tag", "vee"}
}
\value{
no value returned
}
\description{
\code{setGlobalNodeShape} change the shape of all nodes
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setGlobalNodeShape", graph=g)
  layout(rcy, "cose")
  setGlobalNodeShape(rcy, "barrel")
  }

}
