% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setGlobalNodeSize,RCyjs-method}
\alias{setGlobalNodeSize,RCyjs-method}
\alias{setGlobalNodeSize}
\title{setGlobalNodeSize}
\usage{
\S4method{setGlobalNodeSize}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a numeric, in pixels}
}
\value{
no value returned
}
\description{
\code{setGlobalNodeSize} set all nodes to the same specifed size, in pixels
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setGlobalNodesSize", graph=g)
  layout(rcy, "cose")
  setGlobalNodeSize(rcy, 80)
  }

}
