% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setGlobalNodeWidth,RCyjs-method}
\alias{setGlobalNodeWidth,RCyjs-method}
\alias{setGlobalNodeWidth}
\title{setGlobalNodeWidth}
\usage{
\S4method{setGlobalNodeWidth}{RCyjs}(obj, newValue)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{newValue}{a numeric, in pixels}
}
\value{
no value returned
}
\description{
\code{setGlobalNodeWidth} set all nodes to the same specifed width, in pixels
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setGlobalNodesWidth", graph=g)
  layout(rcy, "cose")
  setGlobalNodeWidth(rcy, 80)
  }

}
