% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setNodeAttributes,RCyjs-method}
\alias{setNodeAttributes,RCyjs-method}
\alias{setNodeAttributes}
\title{setNodeAttributes}
\usage{
\S4method{setNodeAttributes}{RCyjs}(obj, attribute, nodes, values)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{attribute}{a character string}

\item{nodes}{character strings - node ids}

\item{values}{scalar values, all of one type (all numeric, or all character, or all integer, ...)}
}
\value{
explain what the method returns
}
\description{
\code{setNodeAttributes} put somewhat more detailed description here
}
\details{
multi-line description goes here with
continuations on subsequent lines
if you like
}
\examples{
if(interactive()){
   g <- simpleDemoGraph()
   rcy <- RCyjs(title="rcyjs demo", graph=g)
   layout(rcy, "cose")
   fit(rcy, 100)
   setNodeAttributes(rcy, "lfc", c("A", "B", "C"), c(0, 0, 0))
   redraw(rcy)
   }


}
