% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setNodeBorderColor,RCyjs-method}
\alias{setNodeBorderColor,RCyjs-method}
\alias{setNodeBorderColor}
\title{setNodeBorderColor}
\usage{
\S4method{setNodeBorderColor}{RCyjs}(obj, nodeIDs, newValues)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{nodeIDs}{a character string (one or more)}

\item{newValue}{a legal CSS color name}
}
\value{
no value returned
}
\description{
\code{setNodeBorderColor} set the specified nodes to the specifed node border color
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setNodeBorderColor", graph=g)
  layout(rcy, "cose")
  setNodeBorderColor(rcy, "green")
  }

}
