% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setNodeBorderWidth,RCyjs-method}
\alias{setNodeBorderWidth,RCyjs-method}
\alias{setNodeBorderWidth}
\title{setNodeBorderWidth}
\usage{
\S4method{setNodeBorderWidth}{RCyjs}(obj, nodeIDs, newValues)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{nodeIDs}{a character string (one or more)}

\item{newValues}{numeric, in pixels (one, or as many as there are nodeIDs)}
}
\value{
no value returned
}
\description{
\code{setNodeBorderWidth} set the specified nodes to the same specifed node border width, in pixels
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setNodeBorderWidth", graph=g)
  layout(rcy, "cose")
  setNodeBorderWidth(rcy, 3)
  }

}
