% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setNodeColor,RCyjs-method}
\alias{setNodeColor,RCyjs-method}
\alias{setNodeColor}
\title{setNodeColor}
\usage{
\S4method{setNodeColor}{RCyjs}(obj, nodeIDs, newValues)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{nodeIDs}{a character string (one or more)}

\item{newValues}{a character string, legal CSS color names (one or more)}
}
\value{
no value returned
}
\description{
\code{setNodeColor} set the specified nodes to the specifed color
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setNodeColor", graph=g)
  layout(rcy, "cose")
  setNodeColor(rcy, 80)
  }

}
