% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setNodeFontColor,RCyjs-method}
\alias{setNodeFontColor,RCyjs-method}
\alias{setNodeFontColor}
\title{setNodeFontColor}
\usage{
\S4method{setNodeFontColor}{RCyjs}(obj, nodeIDs, newValues)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{nodeIDs}{a character string (one or more)}

\item{newValue}{a character string, a legal CSS color name}
}
\value{
no value returned
}
\description{
\code{setNodeFontColor} set the specified nodes to the same specifed node font color
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setNodeFontColor", graph=g)
  layout(rcy, "cose")
  setNodeFontColor(rcy, "red")
  }

}
