% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setNodeFontSize,RCyjs-method}
\alias{setNodeFontSize,RCyjs-method}
\alias{setNodeFontSize}
\title{setNodeFontSize}
\usage{
\S4method{setNodeFontSize}{RCyjs}(obj, nodeIDs, newValues)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{nodeIDs}{a character string (one or more)}

\item{newValues}{a numeric, in pixels (one, or as many as there are nodeIDs)}
}
\value{
no value returned
}
\description{
\code{setNodeFontSize} set the specified nodes to the same specifed node font size
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setNodeFontSize", graph=g)
  layout(rcy, "cose")
  setNodeFontSize(rcy, 5)
  }

}
