% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setNodeHeight,RCyjs-method}
\alias{setNodeHeight,RCyjs-method}
\alias{setNodeHeight}
\title{setNodeHeight}
\usage{
\S4method{setNodeHeight}{RCyjs}(obj, nodeIDs, newValues)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{nodeIDs}{a character string (one or more)}

\item{newValue}{a numeric, in pixels (one, or as many as there are nodeIDs)}
}
\value{
no value returned
}
\description{
\code{setNodeHeight} set the specified nodes to the specifed heights, in pixels
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setNodesHeight", graph=g)
  layout(rcy, "cose")
  setNodeHeight(rcy, 80)
  }

}
