% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setNodeImage,RCyjs-method}
\alias{setNodeImage,RCyjs-method}
\alias{setNodeImage}
\title{setNodeImage}
\usage{
\S4method{setNodeImage}{RCyjs}(obj, imageURLs)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{imageURLs}{a vector of character strings}
}
\value{
explain what the method returns
}
\description{
\code{setNodeImage} put somewhat more detailed description here
}
\details{
multi-line description goes here with
continuations on subsequent lines
if you like
}
\examples{
  x <- 3 + 2

}
