% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setNodeLabelAlignment,RCyjs-method}
\alias{setNodeLabelAlignment,RCyjs-method}
\alias{setNodeLabelAlignment}
\title{setNodelLabelAlignment}
\usage{
\S4method{setNodeLabelAlignment}{RCyjs}(obj, horizontal, vertical)
}
\arguments{
\item{p1}{obj an RCyjs instance}

\item{p2}{horizontal character string}

\item{p3}{vertical character string}
}
\value{
explain what the method returns
}
\description{
\code{setNodeLabelAlignment} put somewhat more detailed description here
}
\details{
multi-line description goes here with
continuations on subsequent lines
if you like
}
\examples{
  x <- 3 + 2

}
