% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setNodeShape,RCyjs-method}
\alias{setNodeShape,RCyjs-method}
\alias{setNodeShape}
\title{setNodeShape}
\usage{
\S4method{setNodeShape}{RCyjs}(obj, nodeIDs, newValues)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{nodeIDs}{a character string (one or more)}

\item{newValues}{a character string, one of the legitimate cytoscape.js node shapes}
}
\value{
no value returned
}
\description{
\code{setNodeShape} set the specified nodes to specifed shapes
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setNodeShape", graph=g)
  layout(rcy, "cose")
  setNodeShape(rcy, 80)
  }

}
