% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setNodeSize,RCyjs-method}
\alias{setNodeSize,RCyjs-method}
\alias{setNodeSize}
\title{setNodeSize}
\usage{
\S4method{setNodeSize}{RCyjs}(obj, nodeIDs, newValues)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{nodeIDs}{a character string (one or more)}

\item{newValue}{a numeric, in pixels (one, or as many as there are nodeIDs)}
}
\value{
no value returned
}
\description{
\code{setNodeSize} set the specified nodes to the specifed sizes, in pixels
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setNodesSize", graph=g)
  layout(rcy, "cose")
  setNodeSize(rcy, 80)
  }

}
