% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setNodeWidth,RCyjs-method}
\alias{setNodeWidth,RCyjs-method}
\alias{setNodeWidth}
\title{setNodeWidth}
\usage{
\S4method{setNodeWidth}{RCyjs}(obj, nodeIDs, newValues)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{nodeIDs}{a character string (one or more)}

\item{newValues}{a numeric, in pixels (one, or as many as there are nodeIDs)}
}
\value{
no value returned
}
\description{
\code{setNodeWidth} set the specified nodes to the specifed widths, in pixels
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="setNodesWidth", graph=g)
  layout(rcy, "cose")
  setNodeWidth(rcy, 80)
  }

}
