% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setPosition,RCyjs-method}
\alias{setPosition,RCyjs-method}
\alias{setPosition}
\title{setPosition}
\usage{
\S4method{setPosition}{RCyjs}(obj, tbl.pos)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{tbl.pos}{a data.frame with three columns: id, x, y}
}
\value{
no return value
}
\description{
\code{setPosition} of nodes by their id
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="getPosition", graph=g)
  layout(rcy, "cose")
  tbl.pos <- getPosition(rcy)
    # shift all the nodes to the right
  tbl.pos$x <- tbl.pos$x + 50
  setPosition(rcy, tbl.pos)
  }

}
\seealso{
\code{\link{getPosition}}
}
