% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{sfn,RCyjs-method}
\alias{sfn,RCyjs-method}
\alias{snf}
\title{sfn}
\usage{
\S4method{sfn}{RCyjs}(obj)
}
\arguments{
\item{obj}{an RCyjs instance}
}
\value{
no return value
}
\description{
\code{sfn} an alias for selectFirstNeighborsOfSelectedNodes
}
\examples{
  x <- 3 + 2

}
