% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{showAll,RCyjs-method}
\alias{showAll,RCyjs-method}
\alias{showAll}
\title{showAll}
\usage{
\S4method{showAll}{RCyjs}(obj, which = c("both", "nodes", "edges"))
}
\arguments{
\item{obj}{an RCyjs instance}

\item{which}{a character string, either "nodes", "edges" or "both"}
}
\value{
no return value
}
\description{
\code{showAll} show any hidden objects: nodes, edges, or both
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="rcyjs demo", graph=g)
  layout(rcy, "cose")
  selectNodes(rcy, getNodes(rcy)$id)
  hideSelectedNodes(rcy)
  showAll(rcy, "nodes")
  }

}
