% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{showNodes,RCyjs-method}
\alias{showNodes,RCyjs-method}
\alias{showNodes}
\title{showNodes}
\usage{
\S4method{showNodes}{RCyjs}(obj, nodeIDs)
}
\arguments{
\item{obj}{an RCyjs instance}
}
\value{
no return value
}
\description{
\code{showNodes} show the named nodes from view
}
\examples{
if(interactive()){
   g <- simpleDemoGraph()
   rcy <- RCyjs(title="rcyjs demo", graph=g)
   target <- nodes(g)[1]
   hideNodes(rcy, "A")
   getNodes(rcy, "hidden")
   getNodes(rcy, "visible")
   showNodes(rcy, "A")
   getNodes(rcy, "visible")
   }

}
\seealso{
\code{\link{showAll}}
}
