% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simpleDemoGraph}
\alias{simpleDemoGraph}
\title{simpleDemoGraph}
\usage{
simpleDemoGraph()
}
\value{
a graphNEL with 3 nodes and 3 edges
}
\description{
simpleDemoGraph
}
\examples{
  g <- simpleDemoGraph()

}
