% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{vAlign,RCyjs-method}
\alias{vAlign,RCyjs-method}
\alias{vAlign}
\title{vAlign}
\usage{
\S4method{vAlign}{RCyjs}(obj)
}
\arguments{
\item{obj}{an RCyjs instance}
}
\value{
no return value
}
\description{
\code{vAlign} vertically align selected nodes
}
\details{
The shared x coordinate will be the mean of the x coordinates of selected nodes.
The y coordinates are preserved.
}
\examples{
if(interactive()){
  g <- simpleDemoGraph()
  rcy <- RCyjs(title="rcyjs demo", graph=g)
  layout(rcy, "cose")
  selectNodes(rcy, nodes(g)[1:2])
  vAlign(rcy)
  }

}
