% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{addGraphFromFile,RCyjs-method}
\alias{addGraphFromFile,RCyjs-method}
\alias{addGraphFromFile}
\title{addGraphFromFile}
\usage{
\S4method{addGraphFromFile}{RCyjs}(obj, jsonFileName)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{jsonFileName}{path to the file}
}
\value{
nothin
}
\description{
\code{addGraphFromFile} add graph from specified file, which contains a cytoscape.js JSON graph
}
\details{
More description
}
\examples{
if(interactive()){
  filename <- system.file(package="RCyjs", "extdata", "sampleGraph.json")
  addGraphFromFile(rcy, filename)
  layout(rcy, "cose")
  fit(rcy, 200)
  }

}
