% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{fit,RCyjs-method}
\alias{fit,RCyjs-method}
\alias{fit}
\title{fit}
\usage{
\S4method{fit}{RCyjs}(obj, padding = 30)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{padding}{numeric, in pixels}
}
\value{
no return value
}
\description{
\code{fit} zoom in (or out) to display all nodes in the current graph
}
\examples{
if(interactive()){
   rcy <- RCyjs(title="rcyjs demo", graph=simpleDemoGraph())
   setZoom(rcy, 0.5)   # zoom out
   fit(rcy)
   }
}
