% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{fitSelection,RCyjs-method}
\alias{fitSelection,RCyjs-method}
\alias{fitSelection}
\title{#' fitSelection}
\usage{
\S4method{fitSelection}{RCyjs}(obj, padding = 30)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{padding}{numeric, pixles  some text}
}
\value{
explain what the method returns
}
\description{
\code{fitSelection} put somewhat more detailed description here
}
\details{
multi-line description goes here with
continuations on subsequent lines
if you like
}
\examples{
  x <- 3 + 2

}
