% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{getEdgeCount,RCyjs-method}
\alias{getEdgeCount,RCyjs-method}
\alias{getEdgeCount}
\title{getEdgeCount}
\usage{
\S4method{getEdgeCount}{RCyjs}(obj)
}
\arguments{
\item{obj}{RCyjs instance}
}
\value{
numeric count
}
\description{
\code{getEdgeCount} the number of edges in the current cytoscape.js graph
}
\examples{
if(interactive()){
   rcy <- RCyjs(title="rcyjs demo", graph=simpleDemoGraph())
   getEdgeCount(rcy)
   }

}
