% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{layoutSelectionInGrid,RCyjs-method}
\alias{layoutSelectionInGrid,RCyjs-method}
\alias{layoutSelectionInGrid}
\title{layoutSelectionInGrid}
\usage{
\S4method{layoutSelectionInGrid}{RCyjs}(obj, x, y, w, h)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{x}{numeric  this will be the top left x coordinate of the grid}

\item{y}{numeric  the top right}

\item{w}{numeric  width of the grid}

\item{h}{numeric  height of the grid}
}
\value{
no return value
}
\description{
\code{layoutSelectionInGrid} arrange selected nodes in this region
}
\examples{
if(interactive()){
   g <- simpleDemoGraph()
   rcy <- RCyjs(title="rcyjs demo", graph=g)
   layout(rcy, "cose")
   fit(rcy, 100)
   loadStyleFile(rcy, system.file(package="RCyjs", "extdata", "sampleStyle2.js"));
   selectNodes(rcy, nodes(g))
   layoutSelectionInGrid(rcy, -1000, 10, 100, 400)
   }

}
