% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{loadStyleFile,RCyjs-method}
\alias{loadStyleFile,RCyjs-method}
\alias{loadStyleFile}
\title{loadStyleFile}
\usage{
\S4method{loadStyleFile}{RCyjs}(obj, filename)
}
\arguments{
\item{filename}{contains json in the proper cytoscape.js format}
}
\value{
nothing
}
\description{
\code{loadStyleFile} load a named JSON cytoscape.js style file into the browser
}
\details{
Though we provide access to individual styling rules (see below) we often find
it convenient to express all aspects of a visual style in a single JSON file
See \link{http://js.cytoscape.org/#style}.
}
\examples{
  if(interactive()){
  rcy <- demo()
  filename <- system.file(package="RCyjs", "extdata", "sampleStyle1.js");
  loadStyleFile(rcy, filename)
  }

}
