% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{redraw,RCyjs-method}
\alias{redraw,RCyjs-method}
\alias{redraw}
\title{redraw}
\usage{
\S4method{redraw}{RCyjs}(obj)
}
\arguments{
\item{obj}{an RCyjs instance}
}
\value{
explain what the method returns
}
\description{
\code{redraw} re-render the graph, using the latest style rules and assignements
}
\examples{
if(interactive()){
   g <- simpleDemoGraph()
   rcy <- RCyjs(title="rcyjs demo", graph=g)
   layout(rcy, "cose")
   fit(rcy, 100)
   setNodeAttributes(rcy, "lfc", c("A", "B", "C"), c(0, 0, 0))
   redraw(rcy)
   }

}
