% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{selectNodes,RCyjs-method}
\alias{selectNodes,RCyjs-method}
\alias{selectNodes}
\title{selectNodes}
\usage{
\S4method{selectNodes}{RCyjs}(obj, nodeIDs)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{nodeIDs}{character strings}
}
\value{
no return value
}
\description{
\code{selectNodes} by node id
}
\examples{
if(interactive()){
   rcy <- RCyjs(title="rcyjs demo", graph=simpleDemoGraph())
   selectNodes(rcy, c("A", "B"))
   }

}
