% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setGraph,RCyjs-method}
\alias{setGraph,RCyjs-method}
\alias{setGraph}
\title{setGraph}
\usage{
\S4method{setGraph}{RCyjs}(obj, graph)
}
\arguments{
\item{obj}{RCyjs instance}

\item{graph}{a graphNEL}
}
\value{
nothing
}
\description{
\code{setGraph} Establish a new graph in RCyjs, removing any previous graph
}
\details{
This method will remove any previous graph in the browser, adding
a new one.  Setting visual properties and performing layout must follow.
}
\examples{
if(interactive()){
  sampleGraph <- simpleDemoGraph()
  rcy <- RCyjs(title="rcyjs demo")
  setGraph(rcy, sampleGraph)
  }

}
\seealso{
\code{\link{addGraph}}
}
