% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setNodeLabelRule,RCyjs-method}
\alias{setNodeLabelRule,RCyjs-method}
\alias{setNodeLabelRule}
\title{setNodeLabelRule}
\usage{
\S4method{setNodeLabelRule}{RCyjs}(obj, attribute)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{attribute}{a character string, the node attribute to display as label}
}
\value{
explain what the method returns
}
\description{
\code{setNodeLabelRule} put somewhat more detailed description here
}
\details{
multi-line description goes here with
continuations on subsequent lines
if you like
}
\examples{
if(interactive()){
  g <- createTestGraph(nodeCount=20, edgeCount=20)
  rcy <- RCyjs(title="layouts", graph=g)
  setNodeLabelRule(rcy, "label");
  }

}
