% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{setNodeSizeRule,RCyjs-method}
\alias{setNodeSizeRule,RCyjs-method}
\alias{setNodeSizeRule}
\title{setNodeSizeRule}
\usage{
\S4method{setNodeSizeRule}{RCyjs}(obj, attribute, control.points, node.sizes)
}
\arguments{
\item{obj}{an RCyjs instance}

\item{attribute}{a character string, the node attribute category whose value controls size}

\item{control.points}{a list of values of the attribute}

\item{node.sizes}{the corresponding node size, one specified for each of the control.points}
}
\value{
no return value
}
\description{
\code{setNodeSizeRule} control node size via values of the specified attribute
}
\details{
actual node sizes are interpolated via the specified relationship of control.points node.sizes
}
\examples{
  x <- 3 + 2

}
