% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCyjs-class.R
\docType{methods}
\name{sfn,RCyjs-method}
\alias{sfn,RCyjs-method}
\alias{sfn}
\title{sfn}
\usage{
\S4method{sfn}{RCyjs}(obj)
}
\arguments{
\item{obj}{an RCyjs instance}
}
\value{
no return value
}
\description{
\code{sfn} select first neighbors of the currently selected nodes
}
\examples{
if(interactive()){
   rcy <- RCyjs(title="rcyjs demo", graph=simpleDemoGraph())
   selectNodes(rcy, "A")
   getSelectedNodes(rcy)  # just one
   sfn()
   getSelectedNodes(rcy)  # now three
   }

}
